fix(@dpc-sdp/nuxt-ripple): fixed unpredictable meta tags by moving from hooks to composibles
